#79 propagating refactoring pattern from #66
UnsafeNativeMethods.shared.cs
- table_get_string removed

CoreProvider.shared.cs
- added StrBufferOverflow, StrBufToStr and StrAllocateBuffer from UnsafeNativeMethods.shared.cs
- GetValue<T> now handles string buffering
